+++ plugin-gradle/src/main/java/com/diffplug/gradle/spotless/SpotlessTaskImpl.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016-2021 DiffPlug
+ * Copyright 2016-2022 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -46,8 +46,18 @@ public abstract class SpotlessTaskImpl extends SpotlessTask {
 	abstract DirectoryProperty getProjectDir();
 
 	void init(Provider<SpotlessTaskService> service) {
+		taskServiceProvider = service;
+		usesService(service);
 		getTaskService().set(service);
 		getProjectDir().set(getProject().getProjectDir());
+	}
+
+	// this field is stupid, but we need it, see https://github.com/diffplug/spotless/issues/1260
+	private transient Provider<SpotlessTaskService> taskServiceProvider;
+
+	@Internal
+	Provider<SpotlessTaskService> getTaskServiceProvider() {
+		return taskServiceProvider;
 	}
 
 	@Inject